server/data-extract/requirements.txt
PyGithub==1.43.8
PyMySQL==0.9.3
Flask==1.1.1
Flask-RESTful==0.3.7